+++ ClothingBrandDashboard/Controllers/CourseController.cs
@@ -1,6 +1,10 @@
 ﻿using ClothingBrandDashboard.ModelVW;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using NuGet.Common;
+using System.Net.Http;
+using System.Net.Http.Headers;
 
 namespace ClothingBrandDashboard.Controllers
 {
@@ -8,9 +12,15 @@ namespace ClothingBrandDashboard.Controllers
     {
         HttpClient client = new HttpClient();
         public string endpoint = "https://localhost:7108/api/Course";
-        public CourseController()
+        private readonly string? token;
+        public CourseController(IHttpContextAccessor httpContextAccessor)
         {
-            
+            token = httpContextAccessor.HttpContext.Session.GetString("AccessToken");
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
+
+            //var token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy93cy8yMDA1LzA1L2lkZW50aXR5L2NsYWltcy9uYW1lIjoiTW9oYW1lZHJhbWFkYW41NDk2QGdtYWlsLmNvbSIsImh0dHA6Ly9zY2hlbWFzLnhtbHNvYXAub3JnL3dzLzIwMDUvMDUvaWRlbnRpdHkvY2xhaW1zL2VtYWlsYWRkcmVzcyI6Ik1vaGFtZWRyYW1hZGFuNTQ5NkBnbWFpbC5jb20iLCJodHRwOi8vc2NoZW1hcy5taWNyb3NvZnQuY29tL3dzLzIwMDgvMDYvaWRlbnRpdHkvY2xhaW1zL3JvbGUiOiJBZG1pbiIsImh0dHA6Ly9zY2hlbWFzLnhtbHNvYXAub3JnL3dzLzIwMDUvMDUvaWRlbnRpdHkvY2xhaW1zL25hbWVpZGVudGlmaWVyIjoiNWEyOGJhMDQtMzU1Ni00MTBmLWFiMGEtNDQxZTQ4OGFhZjA2IiwiRnVsbE5hbWUiOiJNb2hhbWVkIiwiZXhwIjoxNzI5MzQ4MTM3LCJpc3MiOiJodHRwczovL2xvY2FsaG9zdDo3MTA4LyIsImF1ZCI6Imh0dHA6Ly9sb2NhbGhvc3Q6NDIwMC8ifQ.iISt52ilLbjraMxcKPwFs6jAQmsZCr8uctRFO6tGslA";
+            //client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
         }
         public async Task<IActionResult> Index()
         {
@@ -47,6 +57,13 @@ namespace ClothingBrandDashboard.Controllers
 
                 try
                 {
+
+                    if (string.IsNullOrEmpty(token))
+                    {
+                        // Handle case where the token is not available
+                        return Unauthorized();
+                    }
+
                     // Send the POST request to the API
                     var response = await client.PostAsJsonAsync<Course>(endpoint, course);
 
@@ -75,6 +92,11 @@ namespace ClothingBrandDashboard.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(int? id)
         {
+            if (string.IsNullOrEmpty(token))
+            {
+                // Handle case where the token is not available
+                return Unauthorized();
+            }
             //List<Course> Courses = new List<Course>();
             Course course = await client.GetFromJsonAsync<Course>(endpoint + "/" + id);
             Console.WriteLine(course);
@@ -89,6 +111,11 @@ namespace ClothingBrandDashboard.Controllers
                 if (!ModelState.IsValid)
                     return View(course);
 
+                if (string.IsNullOrEmpty(token))
+                {
+                    // Handle case where the token is not available
+                    return Unauthorized();
+                }
                 var response = await client.PutAsJsonAsync(endpoint + "/" + id, course);
 
                 return RedirectToAction(nameof(Index));
@@ -103,6 +130,11 @@ namespace ClothingBrandDashboard.Controllers
         [HttpDelete]
         public IActionResult Delete(int? id)
         {
+            if (string.IsNullOrEmpty(token))
+            {
+                // Handle case where the token is not available
+                return Unauthorized();
+            }
             var course = client.DeleteAsync(endpoint + "/" + id);
             if (course == null)
             {
